Repeatable error messages for metric documentation checks (#28034)
* Sort all keys before checking them for documentation to make sure missing documentation error messages are produced in a predictable and repeatable order.

This is really helpful when someone is trying to fix multiple issues and wants to make sure a specific issue is fixed before moving one to the next one. Otherwise, the errors are thrown in a random order and every time a new message is presented to the developer making it very confusing if the previous one had been resolved or not.