[Elastic Agent] Add process error handling guidelines (#29152)
We don't have a good place yet to document expectations around services run by Elastic Agent. I'm starting this document to have a place to add more content to it but I expect long term we need to figure out a better place.

This guideline comes out of recent issues we had in Cloud and local setups of fleet-server (https://github.com/elastic/fleet-server/pull/883). We never set clear guidlines on what the expectation is of a service run by Elastic Agent and Elastic Agent itself.

This PR is kick off the discussion.


Co-authored-by: Gil Raphaelli <g@raphaelli.com>