Remove macOS preference panels and code related to DMG (#30097)
This commit removes the build of the macOS preference panels for Filebeat,
Auditbeat, Metricbeat, heartbeat. This artifact was never released
officially and was unsupported. We do not have any tests in place to
assert that the new version of macOS doesn't break the preference panel.

Remove the code related to building DMG images, this was only required
to package the preference panel.

Fixes: #30003