Improve test runner to re-use instances and make provisioners pluggable (#3136)
Refactors the way integration test runner works to provide the following
benefits:

* Instance re-use - Call to any `mage integration:*` target that
requires the usage of a deployed instance will re-use an already
deployed instance. A specific call to `mage integration:clean` is
required to bring down created instances and stacks.
* Copy builds only if required - With instance re-use support the code
has been changed to only copy the Elastic Agent builds to an instance in
the case that the build has changed.
* Add `InstanceProvisioner` interface - This interface provides a
defined interface between the runner and the instance provisioner. This
will allow other instance provisioners to be added to the test runner.
* Add `StackProvisioner` interface - This interface provides a defined
interface between the runner and the stack provisioner. This will allow
other stack provisioners to be add to the test runner.