Clean up runtime manager test initialization (#3150)
The runtime manager uses a mysterious lookup table, `waitReady`, that is
injected into the TLS authentication callbacks to allow a special list
of unregistered components to connect. However this turns out to only
ever be used by the unit tests for a single connection, which itself is
only used to probe whether the RPC server is listening yet. This can be
done more simply by just setting an atomic flag when the server loop
starts, so this PR does that and removes the extra synchronization
baggage.

Has no functional effect except when using the unit test helper
`waitForReady`, and should effectively still be a no-op there (just with
fewer redundant network connections).

## Checklist

- [x] My code follows the style guidelines of this project
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] ~~I have made corresponding changes to the documentation~~
- [ ] ~~I have made corresponding change to the default configuration
files~~
- [ ] ~~I have added tests that prove my fix is effective or that my
feature works~~
- [ ] ~~I have added an entry in `./changelog/fragments` using the
[changelog tool](https://github.com/elastic/elastic-agent#changelog)~~
- [ ] ~~I have added an integration test or an E2E test~~