Log raw events to a separate log file (#4549)
This commit introduces a new logger core, used when collecting logs from sub process, that can be configured through `agent.logging.event_data` and is used to log any message that contains the whole event or could contain any sensitive data. This is accomplished by adding `log.type: event` to the log entry. The logger core is responsible for filtering the log entries and directing them to the correct files.