Fix warning on emitting diagnostics for otel hybrid mode (#6973)
If an Otel hybrid mode configuration doesn't exist, we emit a plain text
message into the config file generated by diagnostics. When we try to
redact secrets from it, we fail to unmarshal it into a map and log a
warning about it.

Instead, only run the redaction after confirming the value is actually
a map.