pkg/component/runtime: use atomic.Pointer rather than locks for current component (#7369)
Locking to obtain write boundaries when obtaining the current component.Component
makes use of a read-write mutex. This protection only protects the access and does
not ensure that the current is not mutated with a logical race. The same semantics
can be obtained more cheaply and without the risk of recursive read-lock deadlock
by using an atomic value. This could either be via atomic.Value or atomic.Pointer;
atomic.Pointer is chosen since it can be type-defined in code, rather than by the
addition of a comment.